Rework tests for command arf-to-graph
